requirements.txt
streamlit==1.31.1
pandas==2.2.1
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.6.1
transformers==4.38.1
matplotlib==3.8.4
sacremoses==0.1.1
biopython==1.83
